Implemented parallel functionality for user-defined number of threads on square matrices of user-defined size.

+++ mmm.h
@@ -16,8 +16,8 @@ double mmm_cell(int i, int j);
 
 typedef struct thread_args {
   int tid;    // the given thread id (0, 1, ...)
-  long start; // start row of final matrix this thread is calculating
-  long end;   // end row of final matrix this thread is calculating
+  int start; // start row of final matrix this thread is calculating
+  int end;   // the row directly after end row of final matrix this thread is calculating
 } thread_args;
 
 #endif /* MMM_H_ */
